add kuhap
